requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
cycler==0.11.0
    # via matplotlib
fonttools==4.28.5
    # via matplotlib
kiwisolver==1.3.2
    # via matplotlib
matplotlib==3.5.1
    # via -r requirements.in
networkx==2.6.3
    # via -r requirements.in
numpy==1.21.5
    # via
    #   -r requirements.in
    #   matplotlib
    #   scipy
openseespy==3.4.0.1
    # via -r requirements.in
openseespylinux==3.4.0.1
    # via openseespy
opsvis==0.96.2
    # via -r requirements.in
packaging==21.3
    # via matplotlib
pillow==8.4.0
    # via matplotlib
pyparsing==3.0.6
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via matplotlib
pyyaml==6.0
    # via -r requirements.in
scipy==1.7.3
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
tqdm==4.62.3
    # via -r requirements.in
